requirements.txt
asgiref==3.7.2
asttokens==2.4.1
cairocffi==1.6.1
CairoSVG==2.7.1
cffi==1.16.0
chardet==5.2.0
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.0
executing==2.0.1
gunicorn==21.2.0
ipdb==0.13.13
ipython==8.19.0
jedi==0.19.1
lxml==5.0.0
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
reportlab==4.0.8
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.2.1
traitlets==5.14.0
wcwidth==0.2.12
webencodings==0.5.1
